util: Fix import loop in band
